+++ io.h
@@ -4,10 +4,9 @@
 
 #include <stdio.h>
 
-int read_fd(FILE *fd, unsigned char *buf, size_t buflen, size_t *written);
-int read_file(const char *filename, unsigned char *buf, size_t buflen, size_t *written);
-int read_file_or_stdin(const char *filename, unsigned char *buf,
-                       size_t buflen, size_t *written);
+int read_fd(FILE *fd, unsigned char *buf, int buflen, int *written);
+int read_file(const char *filename, unsigned char *buf, int buflen, int *written);
+int read_file_or_stdin(const char *filename, unsigned char *buf, int buflen, int *written);
 
 
 #endif /* PROTOVERSE_IO_H */
